requirements.txt
astroid==2.15.6
beautifulsoup4==4.12.2
billiard==4.1.0
blinker==1.6.2
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dill==0.3.7
Flask==2.3.2
gunicorn==20.1.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
music-tag==0.4.3
mutagen==1.46.0
pick==2.2.0
Pillow==10.0.0
platformdirs==3.10.0
pycryptodomex==3.18.0
pylint==2.17.5
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.4.1
tomlkit==0.12.1
tqdm==4.65.0
urllib3==2.0.3
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.15.0
yt-dlp==2023.7.6
